backend/ai-service/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
openai==1.6.1
google-generativeai==0.3.2
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
pydantic==2.5.3
psycopg2-binary==2.9.9
redis==5.0.1
pypdf2==3.0.1
python-multipart==0.0.6
